파일 교체

+++ 실습코드/Day13/src/StringEquals.java
@@ -4,13 +4,13 @@ public class StringEquals {
         String str1 = new String("abc");
         String str2 = new String("abc");
 
-        System.out.println(str1 == str2);  // 두 스트링 인스턴스의 주소 값은 다름
-        System.out.println(str1.equals(str2)); // String 클래스의 equals 메소드가 재정의 됨
+        System.out.println(str1 == str2); // 주소값 비교
+        System.out.println(str1.equals(str2)); // 논리적 동일성 비교
 
-        Integer i1 = new Integer(100); // Integer 보다는 Integer.valueOf 사용
-        Integer i2 = new Integer(100);
+        Integer i1 = new Integer(1000);
+        Integer i2 = new Integer(1000);
 
-        System.out.println(i1 == i2);   // 두 정수 인스턴스의 주소 값은 다름
-        System.out.println(i1.equals(i2)); // Integer 클래스의 equals 메서드가 재정의 됨
+        System.out.println(i1 == i2);
+        System.out.println(i1.equals(i2));
     }
 }
